+++ src/main/java/maintest/Test2.java
@@ -1,11 +1,5 @@
 package maintest;
 
-import sauds.image.tools2.Kernel;
-import sauds.image.tools2.*;
-
-import java.io.File;
-import java.io.IOException;
-
 public class Test2 {
 /*
     public static void main(String[] args) throws IOException {
