+++ src/main/java/sauds/image/tools/Aggregator.java
@@ -1,4 +1,4 @@
-package sauds.image.tools2;
+package sauds.image.tools;
 
 import java.util.ArrayList;
 import java.util.Collections;
