+++ src/main/java/sauds/image/tools/BorderHandling.java
@@ -1,4 +1,4 @@
-package sauds.image.tools2;
+package sauds.image.tools;
 
 public enum BorderHandling {
