+++ src/main/java/sauds/image/tools/WriteableImage.java
@@ -1,4 +1,4 @@
-package sauds.image.tools2;
+package sauds.image.tools;
 
 public interface WriteableImage extends Image {
